# [GO Concurrent] Use flattened slice