Main: `12.04 ms  |  165.52 ms  |  1.43 s`

This PR:
```
Go Concurrent:

        Benchmark 1: ./related_concurrent
        Processing time (w/o IO): 11.08487ms
        Processing time (w/o IO): 12.385202ms
        Processing time (w/o IO): 10.692161ms
        Processing time (w/o IO): 10.65006ms
        Processing time (w/o IO): 11.005468ms
        Processing time (w/o IO): 10.733262ms
        Processing time (w/o IO): 10.573858ms
        Processing time (w/o IO): 10.744562ms
        Processing time (w/o IO): 10.887066ms
        Processing time (w/o IO): 10.704161ms
        Processing time (w/o IO): 10.750162ms
        Processing time (w/o IO): 10.731061ms
        Processing time (w/o IO): 10.704761ms
          Time (mean ± σ):      48.6 ms ±   0.8 ms    [User: 73.9 ms, System: 10.3 ms]
          Range (min … max):    47.2 ms …  50.0 ms    10 runs
         
Go Concurrent:

        Benchmark 1: ./related_concurrent
        Processing time (w/o IO): 147.373591ms
        Processing time (w/o IO): 146.311765ms
        Processing time (w/o IO): 146.052959ms
          Time (mean ± σ):     287.4 ms ±   0.2 ms    [User: 708.3 ms, System: 31.6 ms]
          Range (min … max):   287.2 ms … 287.6 ms    2 runs
         
Go Concurrent:

        Benchmark 1: ./related_concurrent
        Processing time (w/o IO): 1.275699608s
        Processing time (w/o IO): 1.285831829s
        Processing time (w/o IO): 1.268917915s
          Time (mean ± σ):      1.744 s ±  0.012 s    [User: 5.525 s, System: 0.093 s]
          Range (min … max):    1.735 s …  1.752 s    2 runs

```